pyproject.toml

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yb"
version = "0.1.0"
description = "A mini protected object store on a YubiKey device"
authors = [{ name = "Frederic Ruget", email = "fred@atlant.com" }]
dependencies = [
  "click",
  "PyYAML",
  "cryptography"
]

[project.scripts]
yb = "yb.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
fix = true
unsafe-fixes = true
